index.html footer cleanup
Consulted client this morning for feedback and content.  Ongoing development as this grows to a future project that will be implemented.  Client requested social media icons in color.  I found .svg to download at 48px and have included them in the footer of the index.html page and cleaned up the section.